Extract wounds editor.
